+++ backend/config/settings.py
@@ -223,7 +223,7 @@ else:
 
 SPECTACULAR_SETTINGS = {
     "TITLE": "Inventory App API",
-    "VERSION": "0.9.1",
+    "VERSION": "0.9.2",
 }
 
 REST_FRAMEWORK = {