+++ CHANGELOG.md
@@ -7,6 +7,8 @@ il versionamento segue [Semantic Versioning](https://semver.org/lang/it/).
 
 ## [Unreleased]
 
+## [0.9.2] - 2026-09-09
+
 ### Changed
 - L'indirizzo mittente di default delle email è ora `noreply@biotron.it` (era `archie@biotron.it`).